file moves in preparation of testing
